batch predict
